Drop uuid: filter when loading the task
Loading tasks does not work with 2.4.0 due to the following bug.

https://bug.tasktools.org/browse/TW-1452

The reason is that vimwiki-tasks uses uuid: filter when loading
the task, however, that filter is broken in 2.4.0.

However, referencing tasks directly (using the task <uuid> info)
command works with TW 2.2, 2.3 and 2.4 (I did not test older
versions), so changing that should be fairly backwards compatible.